initial commit for lambda-headers-conn-cx
This lambda function is used for my website in order to determine cache
times and to also provide sane defaults for security headers.

All this function does is set headers and sometimes set the response
status code to 301 if the origin responds with a 302 (S3 does this when
you leave the trailing slash off of a directory).

Cache is determined first by status, then by paths (to choose lower
times for files with a high amount of change), and then mimetype to
serve binary files longer than text files. If there are no matches for
any of these, we just won't cache the response.

This should allow me to avoid manually invalidating for most activities.
If my menus change or I want to link to new assets, I will need to
invalidate in order to avoid inconsistencies (though, I still wouldn't
end up with broken links).